backend/requirements/base.txt
# Core
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4

# Database
psycopg[binary]>=3.2
dj-database-url>=2.2

# Config
python-decouple>=3.8

# Production server / static
gunicorn>=22.0
whitenoise>=6.7
